stateserver: Automatically allocate all object IDs
This is a significant change in protocol. Now, the stateserver will
automatically assign object IDs to distributed objects, as opposed to
them being manually assigned by an AI/UD server. This means that
generate message types now include a context, which are sent back
with the object information in a new generate response.